358: adding more endpoints and fixing client module
