+++ lib/greenhouse_io/api.rb
@@ -8,6 +8,10 @@ module GreenhouseIo
       self.class.post(url, options)
     end
 
+    def delete_response(url, options)
+      self.class.delete(url, options)
+    end
+
     def parse_json(response)
       JSON.parse(response.body, symbolize_names: GreenhouseIo.configuration.symbolize_keys)
     end
