+++ lib/greenhouse_io/api/modules/candidates.rb
@@ -1,19 +1,16 @@
-require_relative '../request'
-
 module GreenhouseIo
   module Candidates
-    include GreenhouseIo::Request
-
-    def candidate(id, options = {})
-      get_from_harvest_api("/candidates/#{id}", options)
-    end
 
     def candidates(options = {})
       paginated_get("/candidates", options, 'candidates')
     end
 
+    def candidate(id = nil, options = {})
+      get_from_harvest_api("/candidates#{path_id(id)}", options)
+    end
+
     def candidate_activity_feed(id, options = {})
-      get_from_harvest_api "/candidates/#{id}/activity_feed", options
+      get_from_harvest_api "/candidates#{path_id(id)}/activity_feed", options
     end
 
     def create_candidate(candidate_hash, on_behalf_of)
@@ -24,17 +21,9 @@ module GreenhouseIo
       )
     end
 
-    def create_prospect(prospect_hash, on_behalf_of)
-      post_to_harvest_api(
-        "/prospects",
-        prospect_hash,
-        { 'On-Behalf-Of' => on_behalf_of.to_s }
-      )
-    end
-
     def create_candidate_education(candidate_id, education_hash, on_behalf_of)
       post_to_harvest_api(
-        "/candidates/#{candidate_id}/educations"
+        "/candidates#{path_id(candidate_id)}/educations",
         education_hash,
         { 'On-Behalf-Of' => on_behalf_of.to_s }
       )
@@ -42,7 +31,7 @@ module GreenhouseIo
 
     def create_candidate_note(candidate_id, note_hash, on_behalf_of)
       post_to_harvest_api(
-        "/candidates/#{candidate_id}/activity_feed/notes",
+        "/candidates#{path_id(candidate_id)}/activity_feed/notes",
         note_hash,
         { 'On-Behalf-Of' => on_behalf_of.to_s }
       )
