+++ lib/greenhouse_io/api/modules/offers.rb
@@ -1,8 +1,5 @@
-require_relative '../request'
-
 module GreenhouseIo
   module Offers
-    include GreenhouseIo::Request
 
     def offers(id = nil, options = {})
       get_from_harvest_api "/offers#{path_id(id)}", options
